dance_classification/requirements.txt
imageio==2.9.0
imgaug==0.4.0
numpy==1.18.5
numpy-base==1.18.5
opencv-python==4.3.0.36
pandas==1.0.5
pillow==7.2.0
pip==20.1.1
python==3.6.9
pytorch==1.5.1
scikit-image==0.17.2
scipy==1.5.1
setuptools==47.3.1
shapely==1.7.0
torchvision==0.6.1         
